[opt] update main to ts
